refactor(notifications): extract notification history file handling
Move the on-disk half of NotificationManager into a dedicated module.

src/main/services/infrastructure/notificationHistoryFile.ts now owns the
StoredNotification shape, the notification storage path, the two legacy
filenames that are migrated away from, the legacy-path migration, the
history parser (including the recovery path for a truncated or appended
history file) and the atomic write used by the save chain.

NotificationManager keeps the notification pipeline, the toast throttle,
the avatar and Windows-toast helpers and the class itself, and imports
the four values it still needs from the new module. StoredNotification is
re-exported from NotificationManager so every existing caller and the
infrastructure barrel keep the same public surface.

This is pure motion. The moved code is byte-identical apart from the
`export` keywords the new module boundary requires and the two statements
prettier reflows because `export` pushed them past the print width; no
behaviour, name or signature changed. The new module import deliberately
sits in the relative group next to ./ConfigManager: NOTIFICATIONS_PATH and
the legacy paths call getHomeDir() at module-evaluation time, so hoisting
the import above the @main/* block would change when the home directory is
resolved.

NotificationManager.ts 1526 -> 1378 lines; the new module is 179.
No test was touched: NotificationManager.migration.test.ts drives the
migration and parser end-to-end through new NotificationManager().initialize()
and passes unchanged.
